number passed to `boot_and_test` must be an integer

+++ helper/tests/simple_os_checks.py
@@ -84,8 +84,8 @@ def main(argv):
         # storage and volume storage
         #
         # account for count=1 and odd numbers
-        regular_boot_count = (int(count) / 2) + (int(count) % 2)
-        volume_boot_count = int(count) / 2
+        regular_boot_count = int(int(count) / 2) + (int(count) % 2)
+        volume_boot_count = int(int(count) / 2)
         mojo_os_utils.boot_and_test(novac, neutronc,
                                     image_name=image_name,
                                     flavor_name=flavor_name,
